test(crates): drop simulateExecute tests duplicating the mock suite
The tampered logic, compliance, and aggregation proof cases are covered
by `ProtocolAdapterMock.t.sol`. Keep the one check forge cannot do: a
mock seal minted over arm-risc0's aggregation instance must pass
`simulateExecute`, pinning the Solidity encoding to the Rust one.
